web3_unity_backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
web3==6.19.0
eth-account==0.10.0
python-dotenv==1.0.1
